processing_service/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
redis==6.2.0
pydantic==2.11.7
setuptools==80.9.0
jinja2==3.1.4
pyyaml==6.0.2
python-dotenv==1.0.1
